requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.0
matplotlib>=3.8.0
tqdm>=4.66.0
einops>=0.7.0
PyYAML>=6.0.1
scikit-learn>=1.4.0
scipy>=1.12.0
